dog day w/no lunch

+++ src/deployment/setup/updater/update-server-user.sh
@@ -2,13 +2,39 @@ test `id -u` -ne 0 && echo "Root only. Switch to root or use 'sudo'." && exit 1
 
 USERNAME="updater"
 
+check()
+{
+	test -n "`getent passwd $USERNAME`"
+}
+
+fail()
+{
+	printf "Error: $*\n"
+	exit 1
+}
+
 case $1 in
 	help|usage)
-		echo "Usage: `basename $0` --remove-user" >&2
+		echo "Usage: `basename $0` create | check | remove" >&2
 		exit 1
 		;;
-	--remove-user)
+	remove)
+		check
+		test $? || fail "User not found."
 		deluser --remove-home $USERNAME
+		test $? && echo "User sucessfully removed."
+		exit 0
+		;;
+	create)
+		test $? && fail "User already exists."
+		echo "Adding user updater ..."
+		# and go on ...
+		;;
+	check)
+		check
+		test $? || printf "No "
+		printf "user exists.\n"
+		exit 0
 		;;
 	*)
 		echo "Try 'help'." >&2
@@ -16,14 +42,14 @@ case $1 in
 		;;
 esac
 
+# Adding user
+
 SSH_KEY=~/.ssh/esd_rsa
 ADMIN_PUB_KEY=~/.ssh/labrador_rsa.pub
 
-test -n "`getent passwd $USERNAME`" && printf "User already exists.\n" && exit 1
 test ! -f $SSH_KEY && printf "SSH key $SSH_KEY not found.\n" && exit 1
 test ! -f $ADMIN_PUB_KEY && printf "SSH public key $ADMIN_PUB_KEY not found.\n" && exit 1
 
-echo "Adding user updater ..."
 useradd -g users --create-home --shell /bin/bash $USERNAME
 
 SSH_DIR=/home/$USERNAME/.ssh
@@ -38,5 +64,5 @@ cat $SSH_KEY".pub" > $SSH_DIR/authorized_keys
 cat $ADMIN_PUB_KEY >> $SSH_DIR/authorized_keys
 chown -R updater:users $SSH_DIR
 
-echo "User 'updater' created."
+echo "User $USERNAME created."
 echo "To set the password for this user now, use 'sudo passwd updater'."
